+++ scene/resources/font.cpp
@@ -3033,7 +3033,7 @@ Dictionary FontVariation::get_opentype_features() const {
 }
 
 void FontVariation::set_spacing(TextServer::SpacingType p_spacing, int p_value) {
-	ERR_FAIL_INDEX((int)p_spacing, TextServer::SPACING_MAX);
+	ERR_FAIL_INDEX(p_spacing, TextServer::SPACING_MAX);
 	if (extra_spacing[p_spacing] != p_value) {
 		extra_spacing[p_spacing] = p_value;
 		_invalidate_rids();
@@ -3041,7 +3041,7 @@ void FontVariation::set_spacing(TextServer::SpacingType p_spacing, int p_value) 
 }
 
 int FontVariation::get_spacing(TextServer::SpacingType p_spacing) const {
-	ERR_FAIL_INDEX_V((int)p_spacing, TextServer::SPACING_MAX, 0);
+	ERR_FAIL_INDEX_V(p_spacing, TextServer::SPACING_MAX, 0);
 	return extra_spacing[p_spacing];
 }
